Borers: abstracted common features
Noticeable reduction in duplicates
